Did some code cleanup and moved functions
Summarised some functions, changed star imports to aliased imports. Did
major class cleanup on the modvertexbuffers. Removed limit on number of
polygon sides by force triangulating (as it was doing for quads). Did
the gui split.
