+++ vertexbuffers/mhw.py
@@ -1,14 +1,19 @@
 from ..dbg import dbg
-from ..common.fileOperations import *
-from ..common.constants import *
+from ..common import fileOperations as fileOp
+from ..common import constants as cons
+
 def proxima(value1,value2=0,epsilon=0.0001):
     return (value1 <= (value2+epsilon)) and (value1 >= (value2-epsilon))
 
+def MODVertexBufferSelector(BlockType):
+    MVBF = eval("MODVertexBuffer%08x"%BlockType)
+    return MVBF
+
 def calcBonesAndWeightsArr(cnt,weights,bones):
-    weightArrResult = []
-    boneArrResult = []
-    cwt = []
-    cbn = []
+    #weightArrResult = []
+    #boneArrResult = []
+    #cwt = []
+    #cbn = []
     #for b in range(0,cnt):
     #    if (weights[b]>0) and not proxima(weights[b]):
     #        try:
@@ -24,41 +29,38 @@ def calcBonesAndWeightsArr(cnt,weights,bones):
     return (weights,bones)
 
 def calcBonesAndWeights(cnt,weightVal,weightVal2,bns):
-    global WEIGHT_MULTIPLIER
     wt = []
-    w1 = (weightVal & BIT_LENGTH_10)*WEIGHT_MULTIPLIER
-    w2 = ((weightVal>>10) & BIT_LENGTH_10)*WEIGHT_MULTIPLIER
-    w3 = ((weightVal>>20) & BIT_LENGTH_10)*WEIGHT_MULTIPLIER
+    w1 = (weightVal & cons.BIT_LENGTH_10)*cons.WEIGHT_MULTIPLIER
+    w2 = ((weightVal>>10) & cons.BIT_LENGTH_10)*cons.WEIGHT_MULTIPLIER
+    w3 = ((weightVal>>20) & cons.BIT_LENGTH_10)*cons.WEIGHT_MULTIPLIER
     wt.append(w1)
     wt.append(w2)
     wt.append(w3)
     
     if cnt > 4:
-        wt.append((weightVal2[0]) * WEIGHT_MULTIPLIER2)
-        wt.append((weightVal2[1]) * WEIGHT_MULTIPLIER2)
-        wt.append((weightVal2[2]) * WEIGHT_MULTIPLIER2)
-        wt.append((weightVal2[3]) * WEIGHT_MULTIPLIER2)
+        wt.append((weightVal2[0]) * cons.WEIGHT_MULTIPLIER2)
+        wt.append((weightVal2[1]) * cons.WEIGHT_MULTIPLIER2)
+        wt.append((weightVal2[2]) * cons.WEIGHT_MULTIPLIER2)
+        wt.append((weightVal2[3]) * cons.WEIGHT_MULTIPLIER2)
         wt.append(1.0 - wt[0] - wt[1] - wt[2] - wt[3] - wt[4] - wt[5]- wt[6])
         if wt[7] < 0:
             wt[7] = 0
     else:
         w4 = 1.0 - w1 - w2 - w3
-        wt.append(w4)
-
-    
+        wt.append(w4)    
     return calcBonesAndWeightsArr(cnt,wt,bns)
+
 def basicAppendEmptyVertices(cls,fl,VertexRegionEnd,oldVertexCount,addVertexCount):
     dbg("basicAppendEmptyVertices %08x %d %d" % (VertexRegionEnd,oldVertexCount,addVertexCount))
     if(addVertexCount > 0):
-        InsertEmptyBytes(fl,VertexRegionEnd,cls.getStructSize()*addVertexCount)
+        fileOp.InsertEmptyBytes(fl,VertexRegionEnd,cls.getStructSize()*addVertexCount)
     elif(addVertexCount < 0):
         subVertexCount = 0-addVertexCount
-        DeleteBytes(fl,VertexRegionEnd-subVertexCount*cls.getStructSize(),cls.getStructSize()*subVertexCount)
+        fileOp.DeleteBytes(fl,VertexRegionEnd-subVertexCount*cls.getStructSize(),cls.getStructSize()*subVertexCount)
 
-#IANonSkin1UVColor
-class MODVertexBuffer818904dc:
-    def __init__(self,headerref,vertexcount):
-        dbg("MODVertexBuffer818904dc %d" % vertexcount)
+class MODVertexBuffer():
+    def __init__(self, headerref, vertexcount):
+        dbg("%s %d" % (self.__class__.__name__, vertexcount))
         self.vertarray   = []
         self.uvs         = []
         self.weights     = []
@@ -66,63 +68,56 @@ class MODVertexBuffer818904dc:
         self.normalarray = []
         self.headerref   = headerref
         self.vertexcount = vertexcount
-        for i in range(0,vertexcount):
-            if headerref.bendian:
-                self.vertarray.append([ReadBEFloat(headerref.fl),ReadBEFloat(headerref.fl),ReadBEFloat(headerref.fl)])
-            else:
-                self.vertarray.append([ReadFloat(headerref.fl),ReadFloat(headerref.fl),ReadFloat(headerref.fl)])
-            self.normalarray.append((Read8s(headerref.fl),Read8s(headerref.fl),Read8s(headerref.fl)))
-            ReadByte(headerref.fl)
-            ReadLong(headerref.fl)
-            self.uvs.append((ReadHalfFloat(headerref.fl),1-ReadHalfFloat(headerref.fl)))
-            ReadLong(headerref.fl)
-    @staticmethod
-    def getStructSize():
-        return 4+4+4+1+1+1+1+4+2+2+4
+        self.readOp = fileOp.ReadBEFloat if headerref.bendian else fileOp.ReadFloat
     @staticmethod
     def getUVOFFAfterTangents():
         return 0
+    @classmethod
+    def appendEmptyVertices(cls,fl,VertexRegionEnd,oldVertexCount,addVertexCount):
+        basicAppendEmptyVertices(cls,fl,VertexRegionEnd,oldVertexCount,addVertexCount)
+    def readVertexAndNormals(self, headerref):
+        self.vertarray.append([self.readOp(headerref.fl),self.readOp(headerref.fl),self.readOp(headerref.fl)])
+        self.normalarray.append((fileOp.Read8s(headerref.fl),fileOp.Read8s(headerref.fl),fileOp.Read8s(headerref.fl)))
+        fileOp.ReadByte(headerref.fl)
+        fileOp.ReadLong(headerref.fl)
+        
+#IANonSkin1UVColor
+class MODVertexBuffer818904dc(MODVertexBuffer):
+    def __init__(self,headerref,vertexcount):
+        super().__init__(headerref,vertexcount)
+        for i in range(0,vertexcount):
+            self.vertarray.append([self.readOp(headerref.fl),self.readOp(headerref.fl),self.readOp(headerref.fl)])
+            self.normalarray.append((fileOp.Read8s(headerref.fl),fileOp.Read8s(headerref.fl),fileOp.Read8s(headerref.fl)))
+            fileOp.ReadByte(headerref.fl)
+            fileOp.ReadLong(headerref.fl)
+            self.uvs.append((fileOp.ReadHalfFloat(headerref.fl),1-fileOp.ReadHalfFloat(headerref.fl)))
+            fileOp.ReadLong(headerref.fl)
+    @staticmethod
+    def getStructSize():
+        return 4+4+4+1+1+1+1+4+2+2+4
     @staticmethod
     def getWeightsOFFAfterUVOFF():
         return -1
     @staticmethod
     def getBonesOFFAfterWeightsOFF():
         return -1
-    @classmethod
-    def appendEmptyVertices(cls,fl,VertexRegionEnd,oldVertexCount,addVertexCount):
-        basicAppendEmptyVertices(cls,fl,VertexRegionEnd,oldVertexCount,addVertexCount)
+
 
 #IANonSkin2UVColor
-class MODVertexBuffer0f06033f:
+class MODVertexBuffer0f06033f(MODVertexBuffer):
     def __init__(self,headerref,vertexcount):
-        dbg("MODVertexBuffer0f06033f %d" % vertexcount)
-        self.vertarray   = []
-        self.uvs         = []
-        self.weights     = []
-        self.bones       = []
-        self.normalarray = []
-        self.headerref   = headerref
-        self.vertexcount = vertexcount
+        super().__init__(headerref,vertexcount)
         for i in range(0,vertexcount):
-            if headerref.bendian:
-                self.vertarray.append([ReadBEFloat(headerref.fl),ReadBEFloat(headerref.fl),ReadBEFloat(headerref.fl)])
-            else:
-                self.vertarray.append([ReadFloat(headerref.fl),ReadFloat(headerref.fl),ReadFloat(headerref.fl)])
-            self.normalarray.append((Read8s(headerref.fl),Read8s(headerref.fl),Read8s(headerref.fl)))
-            ReadByte(headerref.fl)
-            ReadLong(headerref.fl)
-            self.uvs.append((ReadHalfFloat(headerref.fl),1-ReadHalfFloat(headerref.fl)))
-            wts = ReadLong(headerref.fl)
-            bns = [ReadByte(headerref.fl),ReadByte(headerref.fl),ReadByte(headerref.fl),ReadByte(headerref.fl)]
+            self.readVertexAndNormals(headerref)
+            self.uvs.append((fileOp.ReadHalfFloat(headerref.fl),1-fileOp.ReadHalfFloat(headerref.fl)))
+            wts = fileOp.ReadLong(headerref.fl)
+            bns = [fileOp.ReadByte(headerref.fl),fileOp.ReadByte(headerref.fl),fileOp.ReadByte(headerref.fl),fileOp.ReadByte(headerref.fl)]
             [weights,bones] = calcBonesAndWeights(4,wts,[],bns)
             self.weights.append(weights)
             self.bones.append(bones)
     @staticmethod
     def getStructSize():
         return 4+4+4+1+1+1+1+4+2+2+4+1+ 1+1+1
-    @staticmethod
-    def getUVOFFAfterTangents():
-        return 0
     @staticmethod
     def getWeightsOFFAfterUVOFF():
         return 0
@@ -131,35 +126,19 @@ class MODVertexBuffer0f06033f:
         return 0
     @staticmethod
     def getBoneMode():
-        return WEIGHTS3_BONES4
-    @classmethod
-    def appendEmptyVertices(cls,fl,VertexRegionEnd,oldVertexCount,addVertexCount):
-        basicAppendEmptyVertices(cls,fl,VertexRegionEnd,oldVertexCount,addVertexCount)
+        return cons.WEIGHTS3_BONES4
 
 #IASkin8wt1UV
-class MODVertexBuffer81f58067:
+class MODVertexBuffer81f58067(MODVertexBuffer):
     def __init__(self,headerref,vertexcount):
-        dbg("MODVertexBuffer81f58067 %d" % vertexcount)
-        self.vertarray   = []
-        self.uvs         = []
-        self.weights     = []
-        self.bones       = []
-        self.normalarray = []
-        self.headerref   = headerref
-        self.vertexcount = vertexcount
+        super().__init__(headerref,vertexcount)
         for i in range(0,vertexcount):
-            if headerref.bendian:
-                self.vertarray.append([ReadBEFloat(headerref.fl),ReadBEFloat(headerref.fl),ReadBEFloat(headerref.fl)])
-            else:
-                self.vertarray.append([ReadFloat(headerref.fl),ReadFloat(headerref.fl),ReadFloat(headerref.fl)])
-            self.normalarray.append((Read8s(headerref.fl),Read8s(headerref.fl),Read8s(headerref.fl)))
-            ReadByte(headerref.fl)
-            ReadLong(headerref.fl)
-            self.uvs.append((ReadHalfFloat(headerref.fl),1-ReadHalfFloat(headerref.fl)))
-            wts = ReadLong(headerref.fl)
-            wts2 = [ReadByte(headerref.fl),ReadByte(headerref.fl),ReadByte(headerref.fl),ReadByte(headerref.fl)]            
-            bns = [ReadByte(headerref.fl),ReadByte(headerref.fl),ReadByte(headerref.fl),ReadByte(headerref.fl),
-                ReadByte(headerref.fl),ReadByte(headerref.fl),ReadByte(headerref.fl),ReadByte(headerref.fl)]
+            self.readVertexAndNormals(headerref)
+            self.uvs.append((fileOp.ReadHalfFloat(headerref.fl),1-fileOp.ReadHalfFloat(headerref.fl)))
+            wts = fileOp.ReadLong(headerref.fl)
+            wts2 = [fileOp.ReadByte(headerref.fl),fileOp.ReadByte(headerref.fl),fileOp.ReadByte(headerref.fl),fileOp.ReadByte(headerref.fl)]            
+            bns = [fileOp.ReadByte(headerref.fl),fileOp.ReadByte(headerref.fl),fileOp.ReadByte(headerref.fl),fileOp.ReadByte(headerref.fl),
+                fileOp.ReadByte(headerref.fl),fileOp.ReadByte(headerref.fl),fileOp.ReadByte(headerref.fl),fileOp.ReadByte(headerref.fl)]
             [weights,bones] = calcBonesAndWeights(8,wts,wts2,bns)
             self.weights.append(weights)
             self.bones.append(bones)
@@ -168,9 +147,6 @@ class MODVertexBuffer81f58067:
     def getStructSize():
         return 4+4+4+1+1+1+1+4+2+2+4+1+1+1+1+1 +1+1+1+1+1+1+1
     @staticmethod
-    def getUVOFFAfterTangents():
-        return 0
-    @staticmethod
     def getWeightsOFFAfterUVOFF():
         return 0
     @staticmethod
@@ -178,35 +154,18 @@ class MODVertexBuffer81f58067:
         return 0
     @staticmethod
     def getBoneMode():
-        return WEIGHTS7_BONES8
-    @classmethod
-    def appendEmptyVertices(cls,fl,VertexRegionEnd,oldVertexCount,addVertexCount):
-        basicAppendEmptyVertices(cls,fl,VertexRegionEnd,oldVertexCount,addVertexCount)
+        return cons.WEIGHTS7_BONES8
 
 #IASkin4wt2UV
-class MODVertexBufferf471fe45:
+class MODVertexBufferf471fe45(MODVertexBuffer):
     def __init__(self,headerref,vertexcount):
-        dbg("MODVertexBufferf471fe45 %d" % vertexcount)
-        self.vertarray   = []
-        self.uvs         = []
-        self.uvs2        = []
-        self.weights     = []
-        self.bones       = []
-        self.normalarray = []
-        self.headerref   = headerref
-        self.vertexcount = vertexcount
+        super().__init__(headerref,vertexcount)
         for i in range(0,vertexcount):
-            if headerref.bendian:
-                self.vertarray.append([ReadBEFloat(headerref.fl),ReadBEFloat(headerref.fl),ReadBEFloat(headerref.fl)])
-            else:
-                self.vertarray.append([ReadFloat(headerref.fl),ReadFloat(headerref.fl),ReadFloat(headerref.fl)])
-            self.normalarray.append((Read8s(headerref.fl),Read8s(headerref.fl),Read8s(headerref.fl)))
-            ReadByte(headerref.fl)
-            ReadLong(headerref.fl)
-            self.uvs.append((ReadHalfFloat(headerref.fl),1-ReadHalfFloat(headerref.fl)))
-            self.uvs2.append((ReadHalfFloat(headerref.fl),1-ReadHalfFloat(headerref.fl)))
-            wts = ReadLong(headerref.fl)
-            bns = [ReadByte(headerref.fl),ReadByte(headerref.fl),ReadByte(headerref.fl),ReadByte(headerref.fl)]
+            self.readVertexAndNormals(headerref)
+            self.uvs.append((fileOp.ReadHalfFloat(headerref.fl),1-fileOp.ReadHalfFloat(headerref.fl)))
+            self.uvs2.append((fileOp.ReadHalfFloat(headerref.fl),1-fileOp.ReadHalfFloat(headerref.fl)))
+            wts = fileOp.ReadLong(headerref.fl)
+            bns = [fileOp.ReadByte(headerref.fl),fileOp.ReadByte(headerref.fl),fileOp.ReadByte(headerref.fl),fileOp.ReadByte(headerref.fl)]
             [weights,bones] = calcBonesAndWeights(4,wts,[],bns)
             self.weights.append(weights)
             self.bones.append(bones)
@@ -214,9 +173,6 @@ class MODVertexBufferf471fe45:
     def getStructSize():
         return 4+4+4+1+1+1+1+4+2+2+2+2+4+1+1+1+1
     @staticmethod
-    def getUVOFFAfterTangents():
-        return 0
-    @staticmethod
     def getWeightsOFFAfterUVOFF():
         return 0
     @staticmethod
@@ -224,48 +180,25 @@ class MODVertexBufferf471fe45:
         return 0
     @staticmethod
     def getBoneMode():
-        return WEIGHTS3_BONES4
-    @classmethod
-    def appendEmptyVertices(cls,fl,VertexRegionEnd,oldVertexCount,addVertexCount):
-        basicAppendEmptyVertices(cls,fl,VertexRegionEnd,oldVertexCount,addVertexCount)
+        return cons.WEIGHTS3_BONES4
 
 #IASkin4wt1UVColor
-class MODVertexBuffer3c730760:
+class MODVertexBuffer3c730760(MODVertexBuffer):
     def __init__(self,headerref,vertexcount):
-        dbg("MODVertexBuffer3c730760 %d" % vertexcount)
-        self.vertarray   = []
-        self.uvs         = []
-        self.weights     = []
-        self.bones       = []
-        self.normalarray = []
-        self.headerref   = headerref
-        self.vertexcount = vertexcount
+        super().__init__()
         for i in range(0,vertexcount):
-            if headerref.bendian:
-                self.vertarray.append([ReadBEFloat(headerref.fl),ReadBEFloat(headerref.fl),ReadBEFloat(headerref.fl)])
-            else:
-                self.vertarray.append([ReadFloat(headerref.fl),ReadFloat(headerref.fl),ReadFloat(headerref.fl)])
-            self.normalarray.append((Read8s(headerref.fl),Read8s(headerref.fl),Read8s(headerref.fl)))
-            ReadByte(headerref.fl)
-            ReadLong(headerref.fl)
-            self.uvs.append((ReadHalfFloat(headerref.fl),1-ReadHalfFloat(headerref.fl)))
-            wts = ReadLong(headerref.fl)
-            bns = [ReadByte(headerref.fl),ReadByte(headerref.fl),ReadByte(headerref.fl),ReadByte(headerref.fl)]
+            self.readVertexAndNormals(headerref)
+            self.uvs.append((fileOp.ReadHalfFloat(headerref.fl),1-fileOp.ReadHalfFloat(headerref.fl)))
+            wts = fileOp.ReadLong(headerref.fl)
+            bns = [fileOp.ReadByte(headerref.fl),fileOp.ReadByte(headerref.fl),fileOp.ReadByte(headerref.fl),fileOp.ReadByte(headerref.fl)]
             [weights,bones] = calcBonesAndWeights(4,wts,[],bns)
             self.weights.append(weights)
             self.bones.append(bones)
-            
-            ReadByte(headerref.fl)
-            
-            ReadByte(headerref.fl)
-            ReadByte(headerref.fl)
-            ReadByte(headerref.fl)
+            for _ in range(4):
+                fileOp.ReadByte(headerref.fl)
     @staticmethod
     def getStructSize():
         return 4+4+4+1+1+1+1+4+2+2+4+1+1+1+1+1+1+1+1
-    @staticmethod
-    def getUVOFFAfterTangents():
-        return 0
     @staticmethod
     def getWeightsOFFAfterUVOFF():
         return 0
@@ -274,39 +207,19 @@ class MODVertexBuffer3c730760:
         return 0      
     @staticmethod
     def getBoneMode():
-        return WEIGHTS3_BONES4
-    @classmethod
-    def appendEmptyVertices(cls,fl,VertexRegionEnd,oldVertexCount,addVertexCount):
-        basicAppendEmptyVertices(cls,fl,VertexRegionEnd,oldVertexCount,addVertexCount)
+        return cons.WEIGHTS3_BONES4
 
 #IASkin4wt2UVColor
-class MODVertexBufferb2fc0083:
+class MODVertexBufferb2fc0083(MODVertexBuffer):
     def __init__(self,headerref,vertexcount):
-        dbg("MODVertexBufferb2fc0083 %d" % vertexcount)
-        raise Exception("ToDo")
-        self.vertarray   = []
-        self.uvs         = []
-        self.weights     = []
-        self.bones       = []
-        self.normalarray = []
-        self.headerref   = headerref
-        self.vertexcount = vertexcount
+        super().__init__()
         for i in range(0,vertexcount):
-            if headerref.bendian:
-                self.vertarray.append([ReadBEFloat(headerref.fl),ReadBEFloat(headerref.fl),ReadBEFloat(headerref.fl)])
-            else:
-                self.vertarray.append([ReadFloat(headerref.fl),ReadFloat(headerref.fl),ReadFloat(headerref.fl)])
-            self.normalarray.append((Read8s(headerref.fl),Read8s(headerref.fl),Read8s(headerref.fl)))
-            ReadByte(headerref.fl)
-            ReadLong(headerref.fl)
-            self.uvs.append((ReadHalfFloat(headerref.fl),1-ReadHalfFloat(headerref.fl)))
-            ReadLong(headerref.fl)
+            self.readVertexAndNormals(headerref)
+            self.uvs.append((fileOp.ReadHalfFloat(headerref.fl),1-fileOp.ReadHalfFloat(headerref.fl)))
+            fileOp.ReadLong(headerref.fl)
     @staticmethod
     def getStructSize():
         return 4+4+4+1+1+1+1+4+2+2+4
-    @staticmethod
-    def getUVOFFAfterTangents():
-        return 0
     @staticmethod
     def getWeightsOFFAfterUVOFF():
         return -1
@@ -315,46 +228,27 @@ class MODVertexBufferb2fc0083:
         return -1
     @staticmethod
     def getBoneMode():
-        return WEIGHTS0_BONES0
-    @classmethod
-    def appendEmptyVertices(cls,fl,VertexRegionEnd,oldVertexCount,addVertexCount):
-        basicAppendEmptyVertices(cls,fl,VertexRegionEnd,oldVertexCount,addVertexCount)
+        return cons.WEIGHTS0_BONES0
 
 #IASkin8wt1UVColor
-class MODVertexBuffer366995a7:
+class MODVertexBuffer366995a7(MODVertexBuffer):
     def __init__(self,headerref,vertexcount):
-        dbg("MODVertexBuffer366995a7 %d" % vertexcount)
-        self.vertarray   = []
-        self.uvs         = []
-        self.weights     = []
-        self.bones       = []
-        self.normalarray = []
-        self.headerref   = headerref
-        self.vertexcount = vertexcount
+        super().__init__()
         for i in range(0,vertexcount):
-            if headerref.bendian:
-                self.vertarray.append([ReadBEFloat(headerref.fl),ReadBEFloat(headerref.fl),ReadBEFloat(headerref.fl)])
-            else:
-                self.vertarray.append([ReadFloat(headerref.fl),ReadFloat(headerref.fl),ReadFloat(headerref.fl)])
-            self.normalarray.append((Read8s(headerref.fl),Read8s(headerref.fl),Read8s(headerref.fl)))
-            ReadByte(headerref.fl)
-            ReadLong(headerref.fl)
-            self.uvs.append((ReadHalfFloat(headerref.fl),1-ReadHalfFloat(headerref.fl)))
-            wts = ReadLong(headerref.fl)
-            wts2 = [ReadByte(headerref.fl),ReadByte(headerref.fl),ReadByte(headerref.fl),ReadByte(headerref.fl)]            
-            bns = [ReadByte(headerref.fl),ReadByte(headerref.fl),ReadByte(headerref.fl),ReadByte(headerref.fl),
-                ReadByte(headerref.fl),ReadByte(headerref.fl),ReadByte(headerref.fl),ReadByte(headerref.fl)]
+            self.readVertexAndNormals(headerref)
+            self.uvs.append((fileOp.ReadHalfFloat(headerref.fl),1-fileOp.ReadHalfFloat(headerref.fl)))
+            wts = fileOp.ReadLong(headerref.fl)
+            wts2 = [fileOp.ReadByte(headerref.fl),fileOp.ReadByte(headerref.fl),fileOp.ReadByte(headerref.fl),fileOp.ReadByte(headerref.fl)]            
+            bns = [fileOp.ReadByte(headerref.fl),fileOp.ReadByte(headerref.fl),fileOp.ReadByte(headerref.fl),fileOp.ReadByte(headerref.fl),
+                fileOp.ReadByte(headerref.fl),fileOp.ReadByte(headerref.fl),fileOp.ReadByte(headerref.fl),fileOp.ReadByte(headerref.fl)]
             [weights,bones] = calcBonesAndWeights(8,wts,wts2,bns)
             self.weights.append(weights)
             self.bones.append(bones)
-            ReadLong(headerref.fl)
+            fileOp.ReadLong(headerref.fl)
     @staticmethod
     def getStructSize():
         return 4+4+4+1+1+1+1+4+2+2+4+1+1+1+1 + 8+4
     @staticmethod
-    def getUVOFFAfterTangents():
-        return 0
-    @staticmethod
     def getWeightsOFFAfterUVOFF():
         return 0
     @staticmethod
@@ -362,55 +256,23 @@ class MODVertexBuffer366995a7:
         return 0
     @staticmethod
     def getBoneMode():
-        return WEIGHTS7_BONES8
-    @classmethod
-    def appendEmptyVertices(cls,fl,VertexRegionEnd,oldVertexCount,addVertexCount):
-        basicAppendEmptyVertices(cls,fl,VertexRegionEnd,oldVertexCount,addVertexCount)
+        return cons.WEIGHTS7_BONES8
 
 #FIXME, not referenced by ShaderPackage.sdf, does this even exist?
-class MODVertexBufferc9690ab8:
+class MODVertexBufferc9690ab8(MODVertexBuffer):
     def __init__(self,headerref,vertexcount):
-        dbg("MODVertexBufferc9690ab8 %d" % vertexcount)
-        raise Exception("ToDo")
-        self.vertarray   = []
-        self.uvs         = []
-        self.weights     = []
-        self.bones       = []
-        self.normalarray = []
-        self.headerref   = headerref
-        self.vertexcount = vertexcount
+        super().__init__()
         for i in range(0,vertexcount):
-            if headerref.bendian:
-                self.vertarray.append([ReadBEFloat(headerref.fl),ReadBEFloat(headerref.fl),ReadBEFloat(headerref.fl)])
-            else:
-                self.vertarray.append([ReadFloat(headerref.fl),ReadFloat(headerref.fl),ReadFloat(headerref.fl)])
-            self.normalarray.append((Read8s(headerref.fl),Read8s(headerref.fl),Read8s(headerref.fl)))
-            ReadByte(headerref.fl)
-            ReadLong(headerref.fl)
-            ReadHalfFloat(headerref.fl)
-            ReadHalfFloat(headerref.fl)
-            ReadHalfFloat(headerref.fl)
-            ReadHalfFloat(headerref.fl)
-            ReadByte(headerref.fl)
-            ReadByte(headerref.fl)
-            ReadByte(headerref.fl)
-            ReadByte(headerref.fl)
+            self.readVertexAndNormals(headerref)
+            for _ in range(4):
+                fileOp.ReadHalfFloat(headerref.fl)
+            for _ in range(12):
+                fileOp.ReadByte(headerref.fl)
             
-            ReadByte(headerref.fl)
-            ReadByte(headerref.fl)
-            ReadByte(headerref.fl)
-            ReadByte(headerref.fl)
-            ReadByte(headerref.fl)
-            ReadByte(headerref.fl)
-            ReadByte(headerref.fl)
-            ReadByte(headerref.fl)
     @staticmethod
     def getStructSize():
         return 4+4+4+1+1+1+1+4+2+2+2+2+1+1+1+1 +1+1+1+1+1+1+1+1
     @staticmethod
-    def getUVOFFAfterTangents():
-        return 0
-    @staticmethod
     def getWeightsOFFAfterUVOFF():
         return -1
     @staticmethod
@@ -418,40 +280,21 @@ class MODVertexBufferc9690ab8:
         return -1
     @staticmethod
     def getBoneMode():
-        return WEIGHTS7_BONES8
-    @classmethod
-    def appendEmptyVertices(cls,fl,VertexRegionEnd,oldVertexCount,addVertexCount):
-        basicAppendEmptyVertices(cls,fl,VertexRegionEnd,oldVertexCount,addVertexCount)
+        return cons.WEIGHTS7_BONES8
 
 #FIXME, not referenced by ShaderPackage.sdf, does this even exist?
-class MODVertexBuffer5e7f202d:
+class MODVertexBuffer5e7f202d(MODVertexBuffer):
     def __init__(self,headerref,vertexcount):
-        dbg("MODVertexBuffer5e7f202d %d" % vertexcount)
+        super().__init__()
         raise Exception("ToDo")
-        self.vertarray   = []
-        self.uvs         = []
-        self.weights     = []
-        self.bones       = []
-        self.normalarray = []
-        self.headerref   = headerref
-        self.vertexcount = vertexcount
         for i in range(0,vertexcount):
-            if headerref.bendian:
-                self.vertarray.append([ReadBEFloat(headerref.fl),ReadBEFloat(headerref.fl),ReadBEFloat(headerref.fl)])
-            else:
-                self.vertarray.append([ReadFloat(headerref.fl),ReadFloat(headerref.fl),ReadFloat(headerref.fl)])
-            self.normalarray.append((Read8s(headerref.fl),Read8s(headerref.fl),Read8s(headerref.fl)))
-            ReadByte(headerref.fl)
-            ReadLong(headerref.fl)
-            self.uvs.append((ReadHalfFloat(headerref.fl),1-ReadHalfFloat(headerref.fl)))
-            ReadLong(headerref.fl)
+            self.readVertexAndNormals(headerref)
+            self.uvs.append((fileOp.ReadHalfFloat(headerref.fl),1-fileOp.ReadHalfFloat(headerref.fl)))
+            fileOp.ReadLong(headerref.fl)
     @staticmethod
     def getStructSize():
         return 4+4+4+1+1+1+1+4+2+2+4
     @staticmethod
-    def getUVOFFAfterTangents():
-        return 0
-    @staticmethod
     def getWeightsOFFAfterUVOFF():
         return -1
     @staticmethod
@@ -459,38 +302,19 @@ class MODVertexBuffer5e7f202d:
         return -1
     @staticmethod
     def getBoneMode():
-        return WEIGHTS0_BONES0
-    @classmethod
-    def appendEmptyVertices(cls,fl,VertexRegionEnd,oldVertexCount,addVertexCount):
-        basicAppendEmptyVertices(cls,fl,VertexRegionEnd,oldVertexCount,addVertexCount)
+        return cons.WEIGHTS0_BONES0
 
 #FIXME, not referenced by ShaderPackage.sdf, does this even exist?
-class MODVertexBufferd829702c:
+class MODVertexBufferd829702c(MODVertexBuffer):
     def __init__(self,headerref,vertexcount):
-        dbg("MODVertexBufferd829702c %d" % vertexcount)
-        self.vertarray   = []
-        self.uvs         = []
-        self.weights     = []
-        self.bones       = []
-        self.normalarray = []
-        self.headerref   = headerref
-        self.vertexcount = vertexcount
+        super().__init__()
         for i in range(0,vertexcount):
-            if headerref.bendian:
-                self.vertarray.append([ReadBEFloat(headerref.fl),ReadBEFloat(headerref.fl),ReadBEFloat(headerref.fl)])
-            else:
-                self.vertarray.append([ReadFloat(headerref.fl),ReadFloat(headerref.fl),ReadFloat(headerref.fl)])
-            self.normalarray.append((Read8s(headerref.fl),Read8s(headerref.fl),Read8s(headerref.fl)))
-            ReadByte(headerref.fl)
-            ReadLong(headerref.fl)
-            self.uvs.append((ReadHalfFloat(headerref.fl),1-ReadHalfFloat(headerref.fl)))
+            self.readVertexAndNormals(headerref)
+            self.uvs.append((fileOp.ReadHalfFloat(headerref.fl),1-fileOp.ReadHalfFloat(headerref.fl)))
     @staticmethod
     def getStructSize():
         return 4+4+4+1+1+1+1+4+2+2
     @staticmethod
-    def getUVOFFAfterTangents():
-        return 0
-    @staticmethod
     def getWeightsOFFAfterUVOFF():
         return -1
     @staticmethod
@@ -498,53 +322,30 @@ class MODVertexBufferd829702c:
         return -1
     @staticmethod
     def getBoneMode():
-        return WEIGHTS0_BONES0  
-    @classmethod
-    def appendEmptyVertices(cls,fl,VertexRegionEnd,oldVertexCount,addVertexCount):
-        basicAppendEmptyVertices(cls,fl,VertexRegionEnd,oldVertexCount,addVertexCount)
+        return cons.WEIGHTS0_BONES0  
 
 #IASkin8wt2UVColor
-class MODVertexBufferb8e69244:
+class MODVertexBufferb8e69244(MODVertexBuffer):
     def __init__(self,headerref,vertexcount):
-        dbg("MODVertexBufferd829702c %d" % vertexcount)
-        self.vertarray   = []
-        self.uvs         = []
-        self.weights     = []
-        self.bones       = []
-        self.normalarray = []
-        self.headerref   = headerref
-        self.vertexcount = vertexcount
+        super().__init__()
         for i in range(0,vertexcount):
-            if headerref.bendian:
-                self.vertarray.append([ReadBEFloat(headerref.fl),ReadBEFloat(headerref.fl),ReadBEFloat(headerref.fl)])
-            else:
-                self.vertarray.append([ReadFloat(headerref.fl),ReadFloat(headerref.fl),ReadFloat(headerref.fl)])
-            self.normalarray.append((Read8s(headerref.fl),Read8s(headerref.fl),Read8s(headerref.fl)))
-            ReadByte(headerref.fl)
-            ReadLong(headerref.fl)
-            self.uvs.append((ReadHalfFloat(headerref.fl),1-ReadHalfFloat(headerref.fl)))
-            wts = ReadLong(headerref.fl)
-            wts2 = [ReadByte(headerref.fl),ReadByte(headerref.fl),ReadByte(headerref.fl),ReadByte(headerref.fl)]            
-            bns = [ReadByte(headerref.fl),ReadByte(headerref.fl),ReadByte(headerref.fl),ReadByte(headerref.fl),
-                ReadByte(headerref.fl),ReadByte(headerref.fl),ReadByte(headerref.fl),ReadByte(headerref.fl)]
+            self.readVertexAndNormals(headerref)
+            self.uvs.append((fileOp.ReadHalfFloat(headerref.fl),1-fileOp.ReadHalfFloat(headerref.fl)))
+            wts = fileOp.ReadLong(headerref.fl)
+            wts2 = [fileOp.ReadByte(headerref.fl),fileOp.ReadByte(headerref.fl),fileOp.ReadByte(headerref.fl),fileOp.ReadByte(headerref.fl)]            
+            bns = [fileOp.ReadByte(headerref.fl),fileOp.ReadByte(headerref.fl),fileOp.ReadByte(headerref.fl),fileOp.ReadByte(headerref.fl),
+                fileOp.ReadByte(headerref.fl),fileOp.ReadByte(headerref.fl),fileOp.ReadByte(headerref.fl),fileOp.ReadByte(headerref.fl)]
             [weights,bones] = calcBonesAndWeights(8,wts,wts2,bns)
             self.weights.append(weights)
             self.bones.append(bones)
-
-            ReadLong(headerref.fl)
-            
-            ReadByte(headerref.fl)
-            ReadByte(headerref.fl)
-            ReadByte(headerref.fl)
-            ReadByte(headerref.fl)
+            fileOp.ReadLong(headerref.fl)
+            for _ in range(4):
+                fileOp.ReadByte(headerref.fl)
             
     @staticmethod
     def getStructSize():
         return 4+4+4+1+1+1+1+4+2+2+4+1+1+1+1+1*8+4 +1+1+1+1
     @staticmethod
-    def getUVOFFAfterTangents():
-        return 0
-    @staticmethod
     def getWeightsOFFAfterUVOFF():
         return 0
     @staticmethod
@@ -552,10 +353,7 @@ class MODVertexBufferb8e69244:
         return 0
     @staticmethod
     def getBoneMode():
-        return WEIGHTS7_BONES8
-    @classmethod
-    def appendEmptyVertices(cls,fl,VertexRegionEnd,oldVertexCount,addVertexCount):
-        basicAppendEmptyVertices(cls,fl,VertexRegionEnd,oldVertexCount,addVertexCount)
+        return cons.WEIGHTS7_BONES8
 
 #FIXME IANonSkin2UV    != NonExisting
 MODVertexBuffera5104ca0 = MODVertexBuffer5e7f202d
